+++ LMeter/Config/AboutPage.cs
@@ -7,7 +7,7 @@ namespace LMeter.Config
     public class AboutPage : IConfigPage
     {
 
-        public string Name => "About";
+        public string Name => "Changelog";
 
         public IConfigPage GetDefault() => new AboutPage();