+++ LMeter/Meter/MeterWindow.cs
@@ -7,25 +7,24 @@ using LMeter.Config;
 using LMeter.Helpers;
 using LMeter.ACT;
 using Newtonsoft.Json;
-using System.Globalization;
 
 namespace LMeter.Meter
 {
     public class MeterWindow : IConfigurable
     {
         [JsonIgnore] public readonly string ID;
-        [JsonIgnore] private bool LastFrameWasUnlocked = false;
-        [JsonIgnore] private bool LastFrameWasDragging = false;
-        [JsonIgnore] private bool LastFrameWasPreview = false;
-        [JsonIgnore] private bool Unlocked = false;
-        [JsonIgnore] private bool Hovered = false;
-        [JsonIgnore] private bool Dragging = false;
-        [JsonIgnore] private bool Locked = false;
-        [JsonIgnore] private ACTEvent? PreviewEvent = null;
-        [JsonIgnore] private int ScrollPosition = 0;
 
-        [JsonIgnore] private DateTime? LastSortedTimestamp = null;
-        [JsonIgnore] private List<Combatant> LastSortedCombatants = new List<Combatant>();
+        [JsonIgnore] private bool _lastFrameWasUnlocked = false;
+        [JsonIgnore] private bool _lastFrameWasDragging = false;
+        [JsonIgnore] private bool _lastFrameWasPreview = false;
+        [JsonIgnore] private bool _unlocked = false;
+        [JsonIgnore] private bool _hovered = false;
+        [JsonIgnore] private bool _dragging = false;
+        [JsonIgnore] private bool _locked = false;
+        [JsonIgnore] private ACTEvent? _previewEvent = null;
+        [JsonIgnore] private int _scrollPosition = 0;
+        [JsonIgnore] private DateTime? _lastSortedTimestamp = null;
+        [JsonIgnore] private List<Combatant> _lastSortedCombatants = new List<Combatant>();
 
         public string Name { get; set; }
 
@@ -42,7 +41,7 @@ namespace LMeter.Meter
         public MeterWindow(string name)
         {
             this.Name = name;
-            this.ID = $"LMeter_{GetType().Name}_{Guid.NewGuid()}";
+            this.ID = $"LMeter_MeterWindow_{Guid.NewGuid()}";
             this.GeneralConfig = new GeneralConfig();
             this.HeaderConfig = new HeaderConfig();
             this.BarConfig = new BarConfig();
@@ -97,18 +96,18 @@ namespace LMeter.Meter
 
         public void Clear()
         {
-            this.LastSortedCombatants = new List<Combatant>();
-            this.LastSortedTimestamp = null;
+            this._lastSortedCombatants = new List<Combatant>();
+            this._lastSortedTimestamp = null;
         }
         
         // Dont ask
         protected void UpdateDragData(Vector2 pos, Vector2 size, bool locked)
         {
-            this.Unlocked = !locked;
-            this.Hovered = ImGui.IsMouseHoveringRect(pos, pos + size);
-            this.Dragging = this.LastFrameWasDragging && ImGui.IsMouseDown(ImGuiMouseButton.Left);
-            this.Locked = (this.Unlocked && !this.LastFrameWasUnlocked || !this.Hovered) && !this.Dragging;
-            this.LastFrameWasDragging = this.Hovered || this.Dragging;
+            this._unlocked = !locked;
+            this._hovered = ImGui.IsMouseHoveringRect(pos, pos + size);
+            this._dragging = this._lastFrameWasDragging && ImGui.IsMouseDown(ImGuiMouseButton.Left);
+            this._locked = (this._unlocked && !this._lastFrameWasUnlocked || !this._hovered) && !this._dragging;
+            this._lastFrameWasDragging = this._hovered || this._dragging;
         }
 
         public void Draw(Vector2 pos)
@@ -123,16 +122,16 @@ namespace LMeter.Meter
 
             if (ImGui.IsMouseHoveringRect(localPos, localPos + size))
             {
-                this.ScrollPosition -= (int)ImGui.GetIO().MouseWheel;
+                this._scrollPosition -= (int)ImGui.GetIO().MouseWheel;
             }
 
             this.UpdateDragData(localPos, size, this.GeneralConfig.Lock);
-            bool needsInput = this.Unlocked || !this.GeneralConfig.ClickThrough;
-            DrawHelpers.DrawInWindow($"##{this.ID}", localPos, size, needsInput, this.Locked || this.GeneralConfig.Lock, (drawList) =>
+            bool needsInput = this._unlocked || !this.GeneralConfig.ClickThrough;
+            DrawHelpers.DrawInWindow($"##{this.ID}", localPos, size, needsInput, this._locked || this.GeneralConfig.Lock, (drawList) =>
             {
-                if (this.Unlocked)
+                if (this._unlocked)
                 {
-                    if (this.LastFrameWasDragging)
+                    if (this._lastFrameWasDragging)
                     {
                         localPos = ImGui.GetWindowPos();
                         this.GeneralConfig.Position = localPos - pos;
@@ -154,12 +153,12 @@ namespace LMeter.Meter
                     size -= Vector2.One * this.GeneralConfig.BorderThickness * 2;
                 }
 
-                if (this.GeneralConfig.Preview && !this.LastFrameWasPreview)
+                if (this.GeneralConfig.Preview && !this._lastFrameWasPreview)
                 {
-                    this.PreviewEvent = ACTEvent.GetTestData();
+                    this._previewEvent = ACTEvent.GetTestData();
                 }
 
-                ACTEvent? actEvent = this.GeneralConfig.Preview ? this.PreviewEvent : ACTClient.GetLastEvent();
+                ACTEvent? actEvent = this.GeneralConfig.Preview ? this._previewEvent : ACTClient.GetLastEvent();
                 localPos = this.HeaderConfig.DrawHeader(localPos, size, actEvent?.Encounter, drawList);
                 size = size.AddY(-this.HeaderConfig.HeaderHeight);
 
@@ -168,13 +167,13 @@ namespace LMeter.Meter
                 this.DrawBars(drawList, localPos, size, actEvent);
             });
 
-            this.LastFrameWasUnlocked = this.Unlocked;
-            this.LastFrameWasPreview = this.GeneralConfig.Preview;
+            this._lastFrameWasUnlocked = this._unlocked;
+            this._lastFrameWasPreview = this.GeneralConfig.Preview;
         }
 
         private void DrawBars(ImDrawListPtr drawList, Vector2 localPos, Vector2 size, ACTEvent? actEvent)
         {                
-            if (actEvent is not null && actEvent.Combatants.Any())
+            if (actEvent?.Combatants is not null && actEvent.Combatants.Any())
             {
                 List<Combatant> sortedCombatants = this.GetSortedCombatants(actEvent, this.GeneralConfig.DataType);
                 
@@ -189,8 +188,8 @@ namespace LMeter.Meter
                 int i = 0;
                 if (sortedCombatants.Count > this.BarConfig.BarCount)
                 {
-                    i = Math.Clamp(this.ScrollPosition, 0, sortedCombatants.Count - this.BarConfig.BarCount);
-                    this.ScrollPosition = i;
+                    i = Math.Clamp(this._scrollPosition, 0, sortedCombatants.Count - this.BarConfig.BarCount);
+                    this._scrollPosition = i;
                 }
 
                 int maxIndex = Math.Min(i + this.BarConfig.BarCount, sortedCombatants.Count);
@@ -209,14 +208,7 @@ namespace LMeter.Meter
                     ConfigColor barColor;
                     if (this.BarConfig.UseJobColor)
                     {
-                        if (Enum.TryParse<Job>(combatant.Job, true, out Job job))
-                        {
-                            barColor = this.BarColorsConfig.GetColor(job);
-                        }
-                        else
-                        {
-                            barColor = this.BarColorsConfig.UKNColor;
-                        }
+                        barColor = this.BarColorsConfig.GetColor(combatant.Job);
                     }
                     else
                     {
@@ -230,11 +222,12 @@ namespace LMeter.Meter
 
         private List<Combatant> GetSortedCombatants(ACTEvent actEvent, MeterDataType dataType)
         {
-            if (this.LastSortedTimestamp.HasValue &&
-                this.LastSortedTimestamp.Value == actEvent.Timestamp &&
+            if (actEvent.Combatants is null ||
+                this._lastSortedTimestamp.HasValue &&
+                this._lastSortedTimestamp.Value == actEvent.Timestamp &&
                 !this.GeneralConfig.Preview)
             {
-                return this.LastSortedCombatants;
+                return this._lastSortedCombatants;
             }
 
             List<Combatant> sortedCombatants = actEvent.Combatants.Values.ToList();
@@ -260,8 +253,8 @@ namespace LMeter.Meter
                 return (int)(yFloat - xFloat);
             });
 
-            this.LastSortedTimestamp = actEvent.Timestamp;
-            this.LastSortedCombatants = sortedCombatants;
+            this._lastSortedTimestamp = actEvent.Timestamp;
+            this._lastSortedCombatants = sortedCombatants;
             return sortedCombatants;
         }
     }